requirements.txt
asgiref==3.3.0
boto3==1.16.63
botocore==1.19.63
coverage==5.3.1
dj-database-url==0.5.0
Django==3.0.5
django-admin-interface==0.14.1
django-base64field==1.0
django-colorfield==0.3.2
django-cors-headers==3.5.0
django-cors-middleware==1.5.0
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
djangorestframework==3.12.2
djongo==1.3.3
drf-extra-fields==3.0.4
gunicorn==20.0.4
jmespath==0.10.0
Markdown==3.3.3
mysqlclient==2.0.3
Pillow==8.0.1
psycopg2==2.8.6
pymongo==3.11.2
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
s3transfer==0.3.4
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.2.4
urllib3==1.26.3
whitenoise==5.2.0
